Fix showModal bug on catalog
